--- src/Ideas.metadata.json
@@ -1,11 +1,0 @@
-{
-  "description": "A place for ideas, where they are proposed, discussed, validated, sponsored, and implemented.",
-  "image": {
-    "ipfs_cid": "bafkreifynesvgrnghleag5c7vgshkp7h4uegzwdhjyjqqans4t3hdpuhii"
-  },
-  "name": "Developer Governance Gigs",
-  "tags": {
-    "developer-governance": "",
-    "app": ""
-  }
-}
