+++ src/gigs-board/components/layout/Navbar.jsx
@@ -1,0 +1,84 @@
+/* INCLUDE: "common.jsx" */
+const nearDevGovGigsContractAccountId = props.nearDevGovGigsContractAccountId || (context.widgetSrc ?? 'devgovgigs.near').split('/', 1)[0];
+const nearDevGovGigsWidgetsAccountId = props.nearDevGovGigsWidgetsAccountId || (context.widgetSrc ?? 'devgovgigs.near').split('/', 1)[0];
+
+function widget(widgetName, widgetProps, key) {
+  widgetProps = {
+    ...widgetProps,
+    nearDevGovGigsContractAccountId: props.nearDevGovGigsContractAccountId,
+    nearDevGovGigsWidgetsAccountId: props.nearDevGovGigsWidgetsAccountId,
+  };
+  return <Widget src={`${nearDevGovGigsWidgetsAccountId}/widget/gigs-board.${widgetName}`} props={widgetProps} key={key} />;
+}
+
+function href(widgetName, linkProps) {
+  linkProps = { ...linkProps }
+  if (props.nearDevGovGigsContractAccountId) {
+    linkProps.nearDevGovGigsContractAccountId = props.nearDevGovGigsContractAccountId;
+  }
+  if (props.nearDevGovGigsWidgetsAccountId) {
+    linkProps.nearDevGovGigsWidgetsAccountId = props.nearDevGovGigsWidgetsAccountId;
+  }
+  const linkPropsQuery = Object.entries(linkProps).map(([key, value]) => `${key}=${value}`).join('&');
+  return `#/${nearDevGovGigsWidgetsAccountId}/widget/gigs-board.pages.${widgetName}${linkPropsQuery ? "?" : ""}${linkPropsQuery}`;
+}
+/* END_INCLUDE: "common.jsx" */
+
+return (
+  <div class="card border-secondary">
+    <div class="nav navbar navbar-expand-lg bg-body-tertiary">
+      <div class="container-fluid">
+        <ul class="navbar-nav me-auto mb-2 mb-lg-0">
+          <li class="nav-item ">
+            <a class="nav-link active" href={href("Feed")}>
+              <i class="bi-house-fill"> </i>
+              Home
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link active" href={href("Feed", { recency: "all" })}>
+              <i class="bi-fire"> </i>
+              Recent
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link active" href={href("Feed", { label: "recurrent" })}>
+              <i class="bi-repeat"> </i>
+              Recurrent
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link active" href={href("Boards")}>
+              <i class="bi-kanban"> </i>
+              Boards
+            </a>
+          </li>
+       	  <li class="nav-item">
+            <a
+              class="nav-link active"
+              href={href("Teams")}
+              target="_blank"
+              title="View teams and permissions"
+              role="button"
+            >
+              <i class="bi-people-fill"> </i>
+              Teams
+            </a>
+          </li>
+
+          <li class="nav-item active ms-2">
+            <Typeahead
+              clearButton
+              id="basic-typeahead-single"
+              labelKey="name"
+              onChange={onLabelSelected}
+              options={wrappedLabels}
+              placeholder="Search"
+              defaultSelected={defaultSelectedLabels}
+            />
+          </li>
+        </ul>
+      </div>
+    </div>
+  </div>
+);
