+++ src/gigs-board/components/posts/CompactPost.jsx
@@ -1,6 +1,31 @@
-const ownerId = "devgovgigs.near";
+/* INCLUDE: "common.jsx" */
+const nearDevGovGigsContractAccountId = props.nearDevGovGigsContractAccountId || (context.widgetSrc ?? 'devgovgigs.near').split('/', 1)[0];
+const nearDevGovGigsWidgetsAccountId = props.nearDevGovGigsWidgetsAccountId || (context.widgetSrc ?? 'devgovgigs.near').split('/', 1)[0];
+
+function widget(widgetName, widgetProps, key) {
+  widgetProps = {
+    ...widgetProps,
+    nearDevGovGigsContractAccountId: props.nearDevGovGigsContractAccountId,
+    nearDevGovGigsWidgetsAccountId: props.nearDevGovGigsWidgetsAccountId,
+  };
+  return <Widget src={`${nearDevGovGigsWidgetsAccountId}/widget/gigs-board.${widgetName}`} props={widgetProps} key={key} />;
+}
+
+function href(widgetName, linkProps) {
+  linkProps = { ...linkProps }
+  if (props.nearDevGovGigsContractAccountId) {
+    linkProps.nearDevGovGigsContractAccountId = props.nearDevGovGigsContractAccountId;
+  }
+  if (props.nearDevGovGigsWidgetsAccountId) {
+    linkProps.nearDevGovGigsWidgetsAccountId = props.nearDevGovGigsWidgetsAccountId;
+  }
+  const linkPropsQuery = Object.entries(linkProps).map(([key, value]) => `${key}=${value}`).join('&');
+  return `#/${nearDevGovGigsWidgetsAccountId}/widget/gigs-board.pages.${widgetName}${linkPropsQuery ? "?" : ""}${linkPropsQuery}`;
+}
+/* END_INCLUDE: "common.jsx" */
+
 const postId = props.post.id ?? (props.id ? parseInt(props.id) : 0);
-const post = props.post ?? Near.view(ownerId, "get_post", { post_id: postId });
+const post = props.post ?? Near.view(nearDevGovGigsContractAccountId, "get_post", { post_id: postId });
 if (!post) {
   return <div>Loading ...</div>;
 }
@@ -9,7 +34,7 @@ const snapshot = post.snapshot;
 const shareButton = (
   <a
     class="card-link"
-    href={`https://near.social/#/devgovgigs.near/widget/Ideas?postId=${postId}`}
+    href={href("Feed", { postId })}
     role="button"
     target="_blank"
     title="Open in new tab"
@@ -24,7 +49,7 @@ const header = (
       <div class="row justify-content-between">
         <div class="col-4">
           <a
-            href={`#/mob.near/widget/ProfilePage?accountId=${post.author_id}`}
+            href={`#/neardevgov.near/widget/ProfilePage?accountId=${post.author_id}`}
             className="link-dark text-truncate"
           >
             <Widget
@@ -68,15 +93,14 @@ const renamedPostType =
 
 const postLables = post.snapshot.labels ? (
   <div class="card-title">
-    {post.snapshot.labels.map((label) => {
-      return (
-        <a
-          href={`https://near.social/#/devgovgigs.near/widget/Ideas?label=${label}`}
-        >
-          <span class="badge text-bg-primary me-1">{label}</span>
-        </a>
-      );
-    })}
+    {post.snapshot.labels.map((label) => (
+      <a
+        href={href("Feed", { label })}
+        key={label}
+      >
+        <span class="badge text-bg-primary me-1">{label}</span>
+      </a>
+    ))}
   </div>
 ) : null;
