+++ src/gigs-board/components/teams/LabelsPermissions.jsx
@@ -1,0 +1,66 @@
+/* INCLUDE: "common.jsx" */
+const nearDevGovGigsContractAccountId = props.nearDevGovGigsContractAccountId || (context.widgetSrc ?? 'devgovgigs.near').split('/', 1)[0];
+const nearDevGovGigsWidgetsAccountId = props.nearDevGovGigsWidgetsAccountId || (context.widgetSrc ?? 'devgovgigs.near').split('/', 1)[0];
+
+function widget(widgetName, widgetProps, key) {
+  widgetProps = {
+    ...widgetProps,
+    nearDevGovGigsContractAccountId: props.nearDevGovGigsContractAccountId,
+    nearDevGovGigsWidgetsAccountId: props.nearDevGovGigsWidgetsAccountId,
+  };
+  return <Widget src={`${nearDevGovGigsWidgetsAccountId}/widget/gigs-board.${widgetName}`} props={widgetProps} key={key} />;
+}
+
+function href(widgetName, linkProps) {
+  linkProps = { ...linkProps }
+  if (props.nearDevGovGigsContractAccountId) {
+    linkProps.nearDevGovGigsContractAccountId = props.nearDevGovGigsContractAccountId;
+  }
+  if (props.nearDevGovGigsWidgetsAccountId) {
+    linkProps.nearDevGovGigsWidgetsAccountId = props.nearDevGovGigsWidgetsAccountId;
+  }
+  const linkPropsQuery = Object.entries(linkProps).map(([key, value]) => `${key}=${value}`).join('&');
+  return `#/${nearDevGovGigsWidgetsAccountId}/widget/gigs-board.pages.${widgetName}${linkPropsQuery ? "?" : ""}${linkPropsQuery}`;
+}
+/* END_INCLUDE: "common.jsx" */
+
+const access_info =
+  Near.view(nearDevGovGigsContractAccountId, "get_access_control_info") ?? null;
+
+if (!access_info) {
+  return <div>Loading...</div>;
+}
+
+const rules_list = props.rules_list ?? access_info.rules_list;
+
+const permissionExplainer = (permission) => {
+  if (permission.startsWith("starts-with:")) {
+    let s = permission.substring("starts-with:".length);
+    if (s == "") {
+      return "Any label";
+    } else {
+      return `Labels that start with "${s}"`;
+    }
+  } else {
+    return permission;
+  }
+};
+
+return (
+  <div className="card border-secondary" key="labelpermissions">
+    <div className="card-header">
+      <i class="bi-lock-fill"> </i>
+      <small class="text-muted">Restricted Labels</small>
+    </div>
+    <ul class="list-group list-group-flush">
+      {Object.entries(rules_list).map(([pattern, metadata]) => (
+        <li class="list-group-item" key={pattern}>
+          <span class="badge text-bg-primary" key={`${pattern}-permission`}>
+            {permissionExplainer(pattern)}
+          </span>{" "}
+          {metadata.description}
+        </li>
+      ))}
+    </ul>
+  </div>
+);
