+++ src/gigs-board/pages/Teams.jsx
@@ -1,0 +1,51 @@
+/* INCLUDE: "common.jsx" */
+const nearDevGovGigsContractAccountId = props.nearDevGovGigsContractAccountId || (context.widgetSrc ?? 'devgovgigs.near').split('/', 1)[0];
+const nearDevGovGigsWidgetsAccountId = props.nearDevGovGigsWidgetsAccountId || (context.widgetSrc ?? 'devgovgigs.near').split('/', 1)[0];
+
+function widget(widgetName, widgetProps, key) {
+  widgetProps = {
+    ...widgetProps,
+    nearDevGovGigsContractAccountId: props.nearDevGovGigsContractAccountId,
+    nearDevGovGigsWidgetsAccountId: props.nearDevGovGigsWidgetsAccountId,
+  };
+  return <Widget src={`${nearDevGovGigsWidgetsAccountId}/widget/gigs-board.${widgetName}`} props={widgetProps} key={key} />;
+}
+
+function href(widgetName, linkProps) {
+  linkProps = { ...linkProps }
+  if (props.nearDevGovGigsContractAccountId) {
+    linkProps.nearDevGovGigsContractAccountId = props.nearDevGovGigsContractAccountId;
+  }
+  if (props.nearDevGovGigsWidgetsAccountId) {
+    linkProps.nearDevGovGigsWidgetsAccountId = props.nearDevGovGigsWidgetsAccountId;
+  }
+  const linkPropsQuery = Object.entries(linkProps).map(([key, value]) => `${key}=${value}`).join('&');
+  return `#/${nearDevGovGigsWidgetsAccountId}/widget/gigs-board.pages.${widgetName}${linkPropsQuery ? "?" : ""}${linkPropsQuery}`;
+}
+/* END_INCLUDE: "common.jsx" */
+
+const access_info =
+  Near.view(nearDevGovGigsContractAccountId, "get_access_control_info") ?? null;
+const root_members = Near.view(nearDevGovGigsContractAccountId, "get_root_members") ?? null;
+
+if (!access_info || !root_members) {
+  return <div>Loading...</div>;
+}
+
+const pageContent = (
+  <div>
+    {widget(
+      "components.teams.LabelsPermissions",
+      { rules: access_info.rules_list }
+    )}
+    {Object.keys(root_members).map((member) => (
+      widget(
+        "components.teams.TeamInfo",
+        { member, members_list: access_info.members_list },
+        member
+      )
+    ))}
+  </div>
+);
+
+return widget("components.layout.Page", { children: pageContent })
